requirements.txt
# image processing requirements
opencv-python
pillow

# deep learning requirements
numpy
tensorflow==1.14.0
keras==2.2.4

# python scientific
matplotlib
vmprof